legacy/merchant-server-python/pyproject.toml
[project]
name = "ucp-merchant-server"
version = "0.1.0"
description = "Reference implementation of a UCP Merchant Server using FastAPI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "UCP Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.0",
    "absl-py>=2.3.1",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "fastapi-code-generator>=0.5.4",
    "click==8.1.7",
    "ucp-sdk",
    "shortuuid",
    "httpx>=0.26.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "ruff>=0.14.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
# The relative path is stored here
ucp-sdk = { path = "../../packages/ucp-sdk-python/", editable = true }

[tool.ruff]
line-length = 80
indent-width = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "C4", "SIM", "N", "UP", "D", "PTH", "T20"]
ignore = ["D203", "D213"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
case-sensitive = true
